+++ src/typescript/foundryOverrides/wvToken.ts
@@ -2,6 +2,8 @@ import { getCanvas, getGame } from "../foundryHelpers.js";
 import { getApUse } from "../movement.js";
 import WvActor from "../actor/wvActor.js";
 import { LOG } from "../systemLogger.js";
+import { CONSTANTS } from "../constants.js";
+import { enforceApDragDropSettingName, EnforceApSetting } from "../settings.js";
 
 export default class WvToken extends Token {
   protected override async _onDragLeftDrop(
@@ -9,6 +11,16 @@ export default class WvToken extends Token {
   ): Promise<unknown> {
     // Prepare some stuff and check presence.
     const game = getGame();
+    const setting = game.settings.get(
+      CONSTANTS.systemId,
+      enforceApDragDropSettingName
+    );
+    if (
+      setting === EnforceApSetting.DISABLED ||
+      (setting === EnforceApSetting.PLAYERS && game.user?.isGM)
+    )
+      return super._onDragLeftDrop(event);
+
     const canvas = getCanvas();
     const grid = canvas.grid;
     if (!(grid instanceof GridLayer))